Добавлена ролевая модель

+++ app/views.py
@@ -1,6 +1,7 @@
 import os
 import subprocess
 import uuid
+import logging
 from random import randint
 
 import msal
@@ -9,12 +10,20 @@ from flask import render_template, jsonify, redirect, request, url_for, session,
 from flask_login import login_required, current_user, logout_user, login_user
 
 import config
-from app import app, get_list, _load_cache, _build_msal_app, _save_cache, _build_auth_url, \
-    _get_token_from_cache, db, login_manager
+from app import app, get_machines, _load_cache, _build_msal_app, _save_cache, _build_auth_url, \
+    _get_token_from_cache, db, login_manager, admin_required, get_images, add_vm_to_db, is_owned
 from app.forms import LoginForm
 from app.models import User
-from config import user, hostname, image_folder, high_port, low_port
+from config import DEFAULT_USERNAME, HOSTNAME, image_folder, high_port, low_port
 
+
+logger = logging.getLogger('UI')
+logger.setLevel(logging.DEBUG)
+fh = logging.FileHandler('./app.log')
+fh.setLevel(logging.DEBUG)
+formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+fh.setFormatter(formatter)
+logger.addHandler(fh)
 
 
 @login_manager.user_loader
@@ -36,12 +45,12 @@ def login_passwd():
     form = LoginForm()
     if form.validate_on_submit():
         user = db.session.query(User).filter(User.username == form.username.data).first()
-        print(user)
         if user and user.check_password(form.password.data):
             login_user(user, remember=form.remember.data)
+            logger.info(f"User login via passwd success. Login: '{form.username.data}', passwd: '{form.password.data}'")
             return redirect(url_for('index'))
-        # print(form.password.data, user.password_hash)
         flash("Неправильный логин/пароль", 'error')
+        logger.warning(f"User login via passwd fail. Login: '{form.username.data}', passwd: '{form.password.data}'")
         return redirect(url_for('login_passwd'))
 
     return render_template('login_passwd.html', form=form)
@@ -64,35 +73,30 @@ def authorized():
         session["user"] = result.get("id_token_claims")
         _save_cache(cache)
 
-    # print(session["user"])
-
     users_email = session["user"]["preferred_username"]
     users_name = ' '.join(session["user"]["name"].split(' ')[0:2])
     user = User(
         username=users_name, email=users_email, login_way='msft'
     )
-    print(user)
+
     # Doesn't exist? Add it to the database.
     user_from_db = db.session.query(User).filter(User.username == users_name).first()
     if not user_from_db:
         db.session.add(user)
         db.session.commit()
         user_from_db = db.session.query(User).filter(User.username == users_name).first()
+        logger.warning(f"New user login via MSFT attempt: {session['user']}")
     elif user_from_db.login_way == 'msft':
-        # print("Login w/ MSFT")
-        pass
+        logger.info(f"User login via MSFT attempt: {session['user']}")
     else:
-        print("Login w/ new MSFT")
-        pass
-    # Begin user session by logging the user in
-    # print("111111", current_user)
+        logger.warning(f"User login via MSFT attempt: {session['user']}. This is it's first login via MSFT")
     login_user(user_from_db, remember=True)
-    # print("111111", current_user)
     return redirect(request.args.get('next') or url_for("index"))
 
 
 @app.route("/logout")
 def logout():
+    logger.info(f'User "{current_user.username}" logged out')
     session.clear()  # Wipe out user and its token cache from session
     logout_user()
     return redirect("/")
@@ -116,19 +120,27 @@ def graphcall():
 @app.route("/")
 @login_required
 def index():
-    vms = get_list()
-    print("222", current_user.username)
+    vms = get_machines(current_user.username)
     return render_template('vm_list.html',
-                           vms=vms, vm_user=user, hostname=hostname,)
+                           vms=vms, admin=(current_user.username == config.ADMIN_NAME), hostname=HOSTNAME, )
+
+@app.route("/settings/")
+@admin_required
+def settings():
+    images = get_images()
+    # TODO загрузка и скачивание образа
+    return render_template('settings.html', images=images)
 
 
 @app.route('/launch/<string:id>')
 @login_required
 def launch(id):
-    # TODO проверить, принадлежит ли данная машина текущему пользователю
+    if not is_owned(vm_id=id, username=current_user.username):
+        return redirect("/")
+    # TODO проверить, что свободной паямти хватает
     port = 22003
 
-    vms = get_list()
+    vms = get_machines(user=current_user.username)
     occupied_ports = set([vm['port'] for vm in vms])
     print(occupied_ports)
 
@@ -142,6 +154,8 @@ def launch(id):
         subprocess.check_output(f'vboxmanage startvm --type headless {id}', shell=True)
     except:
         subprocess.check_output(f'vboxmanage modifyvm {id} --natpf1 delete ssh-forwarding', shell=True)
+        subprocess.check_output(f'vboxmanage modifyvm {id} --natpf1 "ssh-forwarding,tcp,,{port},,22"', shell=True)
+        subprocess.check_output(f'vboxmanage startvm --type headless {id}', shell=True)
 
     return redirect('/')
 
@@ -149,7 +163,8 @@ def launch(id):
 @app.route('/stop/<string:id>')
 @login_required
 def stop(id):
-    # TODO проверить, принадлежит ли данная машина текущему пользователю
+    if not is_owned(vm_id=id, username=current_user.username):
+        return redirect("/")
     subprocess.check_output(f'vboxmanage controlvm {id} poweroff', shell=True)
     try:
         subprocess.check_output(f'vboxmanage modifyvm {id} --natpf1 delete ssh-forwarding', shell=True)
@@ -162,21 +177,22 @@ def stop(id):
 @login_required
 def create_vm():
     if request.method == 'POST':
-        # print(dict(request.form))
-        print(os.getcwd(), image_folder, request.form["image"])
+        # TODO проверить уникальность имени
         filename = os.path.abspath(os.path.join(image_folder, request.form["image"]))
         subprocess.check_output(f'vboxmanage import {filename} --vsys 0 --vmname {request.form["name"]}', shell=True)
         subprocess.check_output(f'vboxmanage modifyvm {request.form["name"]} --nic1 nat', shell=True)
-        # TODO добавить запись о машине в БД
-
+        vms = subprocess.check_output('vboxmanage list vms', shell=True).decode('utf-8').split('\n')[:-1]
+        vms = [tuple(vm.split(' ')) for vm in vms]
+        id = [(id[1:-1]) for (name, id) in vms if name[1:-1] == request.form["name"]]
+        if len(id) != 1:
+            return redirect('/')
+        logger.debug(id)
+        add_vm_to_db(image=request.form["image"], name=request.form["name"], user=current_user.username, id_string=id[0])
         return redirect('/')
 
-    # TODO загрузка и скачивание образа
-
+    # TODO проверка, что места на диске хватает
     if not os.path.isdir(image_folder):
         return redirect('/settings/')
-
-    # os.chdir(image_folder)
     images = [file for file in os.listdir(image_folder) if file.endswith('.ova')]
     return render_template('create.html', images=images)
 
@@ -184,16 +200,26 @@ def create_vm():
 @app.route('/delete/<string:id>')
 @login_required
 def delete(id):
-    # TODO проверить, принадлежит ли данная машина текущему пользователю
+    if not is_owned(vm_id=id, username=current_user.username):
+        return redirect("/")
     subprocess.check_output(f'vboxmanage unregistervm {id} --delete', shell=True)
     # TODO удалить запись о машине из БД
     return redirect('/')
 
 
+@app.route('/delete_image/<string:id>')
+@admin_required
+def delete_image(id):
+    # subprocess.check_output(f'vboxmanage unregistervm {id} --delete', shell=True)
+    # TODO удалить запись об образе из БД
+    return redirect('/settings/')
+
+
 @app.route('/monitor/')
+@login_required
 def monitor():
     import psutil
-    vms = [vm for vm in get_list() if vm['running']]
+    vms = [vm for vm in get_machines(user=current_user.username) if vm['running']]
     disk_per = psutil.disk_usage("/")[-1]
     return jsonify(cpu=psutil.cpu_percent(),
                    mem_per=psutil.virtual_memory().percent,
